[Refactor]: S3Service로 코드 리팩터링
