+++ src/main/java/GDSC/FirstProject/config/S3Config.java
@@ -1,4 +1,4 @@
-package GDSC.FirstProject.s3;
+package GDSC.FirstProject.config;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
