--- src/main/java/GDSC/FirstProject/s3/S3Uploader.java
@@ -1,40 +1,0 @@
-package GDSC.FirstProject.s3;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.PutObjectRequest;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-import java.io.File;
-
-@Slf4j
-@RequiredArgsConstructor
-@Service
-public class S3Uploader {
-    private final AmazonS3 amazonS3Client;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-
-    public String upload(String base64, String fileName) {
-        boolean result = Base64ImgDecoder.decode(base64, fileName);
-        System.out.println(result);
-        File uploadFile = new File(fileName);
-
-        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile).withCannedAcl(CannedAccessControlList.PublicRead));
-        removeNewFile(uploadFile);
-
-        return amazonS3Client.getUrl(bucket, fileName).toString();
-    }
-
-    private void removeNewFile(File targetFile) {
-        if (targetFile.delete()) {
-            log.info(targetFile.getName() + "로컬 이미지 파일이 삭제되었습니다.");
-        } else {
-            log.info(targetFile.getName() + "로컬 이미지 파일이 삭제되지 못했습니다.");
-        }
-    }
-}
